Workaround filename variations in subject dump
run and event numbers are encoded in the filename of the images in
the subjec dump. However the structure of this string varies dramatically
through the subject dump, is kept under different keys in the metadata
blob, and is rife with exceptions. Workaround to try and accommodate
any situation encountered in the current subject dump.