Added ability to remove images from panoptes
Need arose to remove all the images in the '0' cluster, which are all
blank images. Added functionality to unlink them from the panoptes subject
set. Also added functionality to add a metadata flag to the structures file
of the affected images, without actually removing them to preserve the data
if necessary. Any image with this flag will not be loaded by the
`Image.load_group` call